Feature: add peek() and peek_mut(); evict when mut ref is given back.
`get_mut()` and `peek_mut()` are different from `get()` or `peek()`,
because it permit user to update the value. And because updating a value
may cause its `measure` to change, therefore when the updated value is
given back to cache, the cache should evict items according to the
measure change.

In this patch, `get_mut()` and `peek_mut()` returns a `PeekMut` instance
that deref to a `&V` and will be responsible to evict items when being
dropped.